+++ src/main/java/com/handwoong/rainbowletter/domain/mail/dto/EmailDto.java
@@ -1,0 +1,5 @@
+package com.handwoong.rainbowletter.domain.mail.dto;
+
+public interface EmailDto {
+    String email();
+}
