+++ src/main/java/com/handwoong/rainbowletter/domain/mail/service/EmailService.java
@@ -1,8 +1,7 @@
-package com.handwoong.rainbowletter.service.mail;
+package com.handwoong.rainbowletter.domain.mail.service;
 
+import com.handwoong.rainbowletter.domain.mail.service.template.EmailTemplateType;
 import jakarta.mail.MessagingException;
-
-import com.handwoong.rainbowletter.service.mail.template.EmailTemplateType;
 
 public interface EmailService {
     void send(final String email, final EmailTemplateType type) throws MessagingException;
