+++ src/main/java/com/handwoong/rainbowletter/domain/mail/service/event/EmailEvent.java
@@ -1,0 +1,9 @@
+package com.handwoong.rainbowletter.domain.mail.service.event;
+
+import com.handwoong.rainbowletter.domain.mail.service.template.EmailTemplateType;
+
+public record EmailEvent(
+        EmailTemplateType type,
+        String email
+) {
+}
