+++ src/main/java/com/handwoong/rainbowletter/domain/mail/service/event/SendEmail.java
@@ -1,11 +1,10 @@
-package com.handwoong.rainbowletter.service.mail.event;
+package com.handwoong.rainbowletter.domain.mail.service.event;
 
+import com.handwoong.rainbowletter.domain.mail.service.template.EmailTemplateType;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
-
-import com.handwoong.rainbowletter.service.mail.template.EmailTemplateType;
 
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
