+++ src/main/java/com/handwoong/rainbowletter/domain/mail/service/template/EmailTemplate.java
@@ -1,6 +1,6 @@
-package com.handwoong.rainbowletter.service.mail.template;
+package com.handwoong.rainbowletter.domain.mail.service.template;
 
-import com.handwoong.rainbowletter.dto.mail.EmailTemplateDto;
+import com.handwoong.rainbowletter.domain.mail.dto.EmailTemplateDto;
 
 public interface EmailTemplate {
     String SUBJECT_FORMAT = "[무지개 편지] %s";
