+++ src/main/java/com/handwoong/rainbowletter/domain/mail/service/template/EmailVerifyTemplate.java
@@ -1,16 +1,14 @@
-package com.handwoong.rainbowletter.service.mail.template;
-
-import org.springframework.stereotype.Component;
-import org.thymeleaf.TemplateEngine;
-import org.thymeleaf.context.Context;
+package com.handwoong.rainbowletter.domain.mail.service.template;
 
 import com.handwoong.rainbowletter.config.PropertiesConfig;
 import com.handwoong.rainbowletter.config.security.GrantType;
 import com.handwoong.rainbowletter.config.security.JwtTokenProvider;
 import com.handwoong.rainbowletter.config.security.TokenResponse;
-import com.handwoong.rainbowletter.dto.mail.EmailTemplateDto;
-
+import com.handwoong.rainbowletter.domain.mail.dto.EmailTemplateDto;
 import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+import org.thymeleaf.TemplateEngine;
+import org.thymeleaf.context.Context;
 
 @Component
 @RequiredArgsConstructor
@@ -33,7 +31,8 @@ public class EmailVerifyTemplate implements EmailTemplate {
     private String createBody(final String email) {
         final TokenResponse tokenResponse =
                 tokenProvider.generateToken(GrantType.PATH_VARIABLE, email, EmailTemplateType.VERIFY.name());
-        final String verifyUrl = propertiesConfig.getClientUrls().get(0) + "/members/verify?token=" + tokenResponse.token();
+        final String verifyUrl =
+                propertiesConfig.getClientUrls().get(0) + "/members/verify?token=" + tokenResponse.token();
         final Context context = new Context();
         context.setVariable("verifyUrl", verifyUrl);
         return templateEngine.process("verify", context);
