+++ src/main/java/com/handwoong/rainbowletter/domain/member/dto/ChangePasswordRequest.java
@@ -1,8 +1,8 @@
 package com.handwoong.rainbowletter.domain.member.dto;
 
-import static com.handwoong.rainbowletter.dto.ValidateMessage.EMPTY_MESSAGE;
-import static com.handwoong.rainbowletter.dto.ValidateMessage.PASSWORD_FORMAT;
-import static com.handwoong.rainbowletter.dto.ValidateMessage.PASSWORD_MESSAGE;
+import static com.handwoong.rainbowletter.util.ValidateMessage.EMPTY_MESSAGE;
+import static com.handwoong.rainbowletter.util.ValidateMessage.PASSWORD_FORMAT;
+import static com.handwoong.rainbowletter.util.ValidateMessage.PASSWORD_MESSAGE;
 
 import jakarta.annotation.Nullable;
 import jakarta.validation.constraints.NotBlank;
