+++ src/main/java/com/handwoong/rainbowletter/domain/member/dto/ChangePhoneNumberRequest.java
@@ -1,8 +1,8 @@
 package com.handwoong.rainbowletter.domain.member.dto;
 
-import static com.handwoong.rainbowletter.dto.ValidateMessage.EMPTY_MESSAGE;
-import static com.handwoong.rainbowletter.dto.ValidateMessage.PHONE_NUMBER_FORMAT;
-import static com.handwoong.rainbowletter.dto.ValidateMessage.PHONE_NUMBER_MESSAGE;
+import static com.handwoong.rainbowletter.util.ValidateMessage.EMPTY_MESSAGE;
+import static com.handwoong.rainbowletter.util.ValidateMessage.PHONE_NUMBER_FORMAT;
+import static com.handwoong.rainbowletter.util.ValidateMessage.PHONE_NUMBER_MESSAGE;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
