+++ src/main/java/com/handwoong/rainbowletter/domain/member/dto/FindPasswordDto.java
@@ -1,9 +1,9 @@
 package com.handwoong.rainbowletter.domain.member.dto;
 
-import static com.handwoong.rainbowletter.dto.ValidateMessage.EMAIL_MESSAGE;
-import static com.handwoong.rainbowletter.dto.ValidateMessage.EMPTY_MESSAGE;
+import static com.handwoong.rainbowletter.util.ValidateMessage.EMAIL_MESSAGE;
+import static com.handwoong.rainbowletter.util.ValidateMessage.EMPTY_MESSAGE;
 
-import com.handwoong.rainbowletter.dto.mail.EmailDto;
+import com.handwoong.rainbowletter.domain.mail.dto.EmailDto;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
