+++ src/main/java/com/handwoong/rainbowletter/domain/member/dto/MemberRegisterRequest.java
@@ -1,9 +1,9 @@
 package com.handwoong.rainbowletter.domain.member.dto;
 
-import static com.handwoong.rainbowletter.dto.ValidateMessage.EMAIL_MESSAGE;
-import static com.handwoong.rainbowletter.dto.ValidateMessage.EMPTY_MESSAGE;
-import static com.handwoong.rainbowletter.dto.ValidateMessage.PASSWORD_FORMAT;
-import static com.handwoong.rainbowletter.dto.ValidateMessage.PASSWORD_MESSAGE;
+import static com.handwoong.rainbowletter.util.ValidateMessage.EMAIL_MESSAGE;
+import static com.handwoong.rainbowletter.util.ValidateMessage.EMPTY_MESSAGE;
+import static com.handwoong.rainbowletter.util.ValidateMessage.PASSWORD_FORMAT;
+import static com.handwoong.rainbowletter.util.ValidateMessage.PASSWORD_MESSAGE;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
