+++ src/main/java/com/handwoong/rainbowletter/domain/member/dto/MemberRegisterResponse.java
@@ -1,7 +1,7 @@
 package com.handwoong.rainbowletter.domain.member.dto;
 
+import com.handwoong.rainbowletter.domain.mail.dto.EmailDto;
 import com.handwoong.rainbowletter.domain.member.model.Member;
-import com.handwoong.rainbowletter.dto.mail.EmailDto;
 
 public record MemberRegisterResponse(
         Long id,
