+++ src/main/java/com/handwoong/rainbowletter/domain/member/service/MemberService.java
@@ -1,6 +1,7 @@
 package com.handwoong.rainbowletter.domain.member.service;
 
 import com.handwoong.rainbowletter.config.security.TokenResponse;
+import com.handwoong.rainbowletter.domain.mail.dto.EmailDto;
 import com.handwoong.rainbowletter.domain.member.dto.ChangePasswordRequest;
 import com.handwoong.rainbowletter.domain.member.dto.ChangePhoneNumberRequest;
 import com.handwoong.rainbowletter.domain.member.dto.FindPasswordDto;
@@ -8,7 +9,6 @@ import com.handwoong.rainbowletter.domain.member.dto.MemberLoginRequest;
 import com.handwoong.rainbowletter.domain.member.dto.MemberRegisterRequest;
 import com.handwoong.rainbowletter.domain.member.dto.MemberRegisterResponse;
 import com.handwoong.rainbowletter.domain.member.dto.MemberResponse;
-import com.handwoong.rainbowletter.dto.mail.EmailDto;
 
 public interface MemberService {
     MemberResponse info(final String email);
