+++ src/main/java/com/handwoong/rainbowletter/domain/member/service/MemberServiceImpl.java
@@ -3,6 +3,9 @@ package com.handwoong.rainbowletter.domain.member.service;
 import com.handwoong.rainbowletter.config.security.GrantType;
 import com.handwoong.rainbowletter.config.security.JwtTokenProvider;
 import com.handwoong.rainbowletter.config.security.TokenResponse;
+import com.handwoong.rainbowletter.domain.mail.dto.EmailDto;
+import com.handwoong.rainbowletter.domain.mail.service.event.SendEmail;
+import com.handwoong.rainbowletter.domain.mail.service.template.EmailTemplateType;
 import com.handwoong.rainbowletter.domain.member.dto.ChangePasswordRequest;
 import com.handwoong.rainbowletter.domain.member.dto.ChangePhoneNumberRequest;
 import com.handwoong.rainbowletter.domain.member.dto.FindPasswordDto;
@@ -13,11 +16,8 @@ import com.handwoong.rainbowletter.domain.member.dto.MemberResponse;
 import com.handwoong.rainbowletter.domain.member.model.Member;
 import com.handwoong.rainbowletter.domain.member.model.MemberStatus;
 import com.handwoong.rainbowletter.domain.member.repository.MemberRepository;
-import com.handwoong.rainbowletter.dto.mail.EmailDto;
 import com.handwoong.rainbowletter.exception.ErrorCode;
 import com.handwoong.rainbowletter.exception.RainbowLetterException;
-import com.handwoong.rainbowletter.service.mail.event.SendEmail;
-import com.handwoong.rainbowletter.service.mail.template.EmailTemplateType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
