--- src/main/java/com/handwoong/rainbowletter/dto/mail/EmailDto.java
@@ -1,5 +1,0 @@
-package com.handwoong.rainbowletter.dto.mail;
-
-public interface EmailDto {
-    String email();
-}
