--- src/main/java/com/handwoong/rainbowletter/service/mail/event/EmailEvent.java
@@ -1,9 +1,0 @@
-package com.handwoong.rainbowletter.service.mail.event;
-
-import com.handwoong.rainbowletter.service.mail.template.EmailTemplateType;
-
-public record EmailEvent(
-        EmailTemplateType type,
-        String email
-) {
-}
