+++ src/test/java/com/handwoong/rainbowletter/domain/member/controller/MemberControllerTest.java
@@ -14,12 +14,12 @@ import com.handwoong.rainbowletter.config.security.GrantType;
 import com.handwoong.rainbowletter.config.security.JwtTokenProvider;
 import com.handwoong.rainbowletter.config.security.TokenResponse;
 import com.handwoong.rainbowletter.domain.ControllerTestProvider;
+import com.handwoong.rainbowletter.domain.mail.service.template.EmailTemplateType;
 import com.handwoong.rainbowletter.domain.member.dto.ChangePasswordRequest;
 import com.handwoong.rainbowletter.domain.member.dto.ChangePhoneNumberRequest;
 import com.handwoong.rainbowletter.domain.member.dto.FindPasswordDto;
 import com.handwoong.rainbowletter.domain.member.dto.MemberLoginRequest;
 import com.handwoong.rainbowletter.domain.member.dto.MemberRegisterRequest;
-import com.handwoong.rainbowletter.service.mail.template.EmailTemplateType;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
